move app widget files to its own package
