+++ app/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
         </activity>
 
         <receiver
-            android:name=".MyAppWidgetReceiver"
+            android:name=".AppWidget.MyAppWidgetReceiver"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
